requirements.txt
django-mptt==0.16.0
django-admin-thumbnails==0.2.8
django-ckeditor==6.7.1
query==0.1.4
Pillow==10.3.0
psycopg2-binary==2.9.9
django-storages
azure-storage-blob
django-environ